data_gen/requirements.txt
faker=8.8.1
python-rapidjson=1.5
optional-faker==1.0.0.post2